requirements.txt
streamlit==1.16.0
numpy==1.23.0 
tensorflow==2.12.0
Pillow==10.0.0
